Created API routes to fetch orders, change delivery status and to create pdf of todays orders also displaying total earning and top selling product

+++ app/api/deleteChanges/route.js
@@ -1,9 +1,16 @@
 import { NextResponse } from "next/server";
 import { deleteProducts } from "@/actions/useractions";
+import { getToken } from "next-auth/jwt";
 
-export async function DELETE(req){
-    let body = await req.json()
-    let data = await deleteProducts(body.deleteIDs)
-    
-    return NextResponse.json({"success": true})
+export async function DELETE(req) {
+    try {
+        const token = await getToken({ req })
+        if (!token || !token.email) throw new Error("Unauthorized");
+        let body = await req.json()
+        let data = await deleteProducts(body.deleteIDs)
+        if (!data.success) throw new Error(data.message)
+        return NextResponse.json({ success: data.success, message: data.message })
+    } catch (error) {
+        return NextResponse.json({ success: false, message: error })
+    }
 }
